+++ lib/savon/soap/request.rb
@@ -12,6 +12,12 @@ module Savon
       # Content-Types by SOAP version.
       ContentType = { 1 => "text/xml;charset=UTF-8", 2 => "application/soap+xml;charset=UTF-8" }
 
+      # Expects an <tt>HTTPI::Request</tt> and a <tt>Savon::SOAP::XML</tt> object
+      # to execute a SOAP request and returns the response.
+      def self.execute(request, soap)
+        new(request, soap).response
+      end
+
       # Expects an <tt>HTTPI::Request</tt> and a <tt>Savon::SOAP::XML</tt> object.
       def initialize(request, soap)
         self.request = setup(request, soap)
@@ -45,6 +51,7 @@ module Savon
 
       # Logs the SOAP request +url+, +headers+ and +body+.
       def log_request(url, headers, body)
+        Savon.log "----"
         Savon.log "SOAP request: #{url}"
         Savon.log headers.map { |key, value| "#{key}: #{value}" }.join(", ")
         Savon.log body
@@ -52,6 +59,7 @@ module Savon
 
       # Logs the SOAP response +code+ and +body+.
       def log_response(code, body)
+        Savon.log "----"
         Savon.log "SOAP response (status #{code}):"
         Savon.log body
       end