+++ lib/savon/soap/xml.rb
@@ -20,31 +20,6 @@ module Savon
         "xmlns:xsi" => "http://www.w3.org/2001/XMLSchema-instance"
       }
 
-      # Converts the given SOAP response +value+ (XML or Hash) into a normalized Hash.
-      def self.to_hash(value)
-        value = parse value unless value.kind_of? Hash
-        value.find_soap_body
-      end
-
-      # Converts a given SOAP response +xml+ to a Hash.
-      def self.parse(xml)
-        Nori.parse(xml)
-      end
-
-      # Expects a SOAP response XML or Hash, traverses it for a given +path+ of Hash keys
-      # and returns the value as an Array. Defaults to return an empty Array in case the
-      # path does not exist or returns nil.
-      def self.to_array(object, *path)
-        hash = object.kind_of?(Hash) ? object : to_hash(object)
-
-        result = path.inject hash do |memo, key|
-          return [] unless memo[key]
-          memo[key]
-        end
-
-        result.kind_of?(Array) ? result.compact : [result].compact
-      end
-
       # Accepts an +endpoint+, an +input+ tag and a SOAP +body+.
       def initialize(endpoint = nil, input = nil, body = nil)
         self.endpoint = endpoint if endpoint
@@ -74,7 +49,7 @@ module Savon
 
       # Returns the SOAP +header+. Defaults to an empty Hash.
       def header
-        @header ||= {}
+        @header ||= Savon.soap_header.nil? ? {} : Savon.soap_header
       end
 
       # Sets the SOAP envelope namespace.