+++ spec/savon/client_spec.rb
@@ -1,72 +1,65 @@
 require "spec_helper"
 
 describe Savon::Client do
-  let(:client) { Savon::Client.new { wsdl.document = Endpoint.wsdl } }
+  let(:client) { Savon::Client.new Endpoint.wsdl }
 
   describe ".new" do
-    context "with a String" do
-      it "should set the WSDL document" do
+    context "called with a String" do
+      it "sets the WSDL document" do
         wsdl = "http://example.com/UserService?wsdl"
         client = Savon::Client.new wsdl
         client.wsdl.instance_variable_get("@document").should == wsdl
       end
     end
 
-    context "with a block expecting one argument" do
-      it "should yield the WSDL object" do
-        Savon::Client.new { |wsdl| wsdl.should be_a(Savon::WSDL::Document) }
+    context "called with a block expecting one argument" do
+      it "yields the client instance" do
+        Savon::Client.new { |client| client.should be_a(Savon::Client) }
       end
     end
 
-    context "with a block expecting two arguments" do
-      it "should yield the WSDL and HTTP objects" do
-        Savon::Client.new do |wsdl, http|
-          wsdl.should be_an(Savon::WSDL::Document)
-          http.should be_an(HTTPI::Request)
-        end
-      end
-    end
-
-    context "with a block expecting three arguments" do
-      it "should yield the WSDL, HTTP and WSSE objects" do
-        Savon::Client.new do |wsdl, http, wsse|
-          wsdl.should be_an(Savon::WSDL::Document)
-          http.should be_an(HTTPI::Request)
-          wsse.should be_an(Savon::WSSE)
-        end
-      end
-    end
-
-    context "with a block expecting no arguments" do
-      it "should let you access the WSDL object" do
+    context "called with a block expecting no arguments" do
+      it "lets you access the WSDL object" do
         Savon::Client.new { wsdl.should be_a(Savon::WSDL::Document) }
       end
 
-      it "should let you access the HTTP object" do
+      it "lets you access the HTTP object" do
         Savon::Client.new { http.should be_an(HTTPI::Request) }
       end
 
-      it "should let you access the WSSE object" do
+      it "lets you access the WSSE object" do
         Savon::Client.new { wsse.should be_a(Savon::WSSE) }
       end
     end
   end
 
   describe "#wsdl" do
-    it "should return the Savon::WSDL::Document" do
+    it "returns the Savon::WSDL::Document" do
       client.wsdl.should be_a(Savon::WSDL::Document)
+    end
+
+    it "memoizes the object" do
+      client.wsdl.should equal(client.wsdl)
     end
   end
 
   describe "#http" do
-    it "should return the HTTPI::Request" do
+    it "returns the HTTPI::Request" do
       client.http.should be_an(HTTPI::Request)
+    end
+
+    it "memoizes the object" do
+      client.http.should equal(client.http)
     end
   end
 
   describe "#wsse" do
-    it "should return the Savon::WSSE object" do
+    it "returns the Savon::WSSE object" do
       client.wsse.should be_a(Savon::WSSE)
+    end
+
+    it "memoizes the object" do
+      client.wsse.should equal(client.wsse)
     end
   end
 
@@ -76,25 +69,26 @@ describe Savon::Client do
       HTTPI.stubs(:post).returns(new_response)
     end
 
-    context "without any arguments" do
-      it "should raise an ArgumentError" do
-        lambda { client.request }.should raise_error(ArgumentError)
+    context "called without any arguments" do
+      it "raises an ArgumentError" do
+        message = "Expected to receive at least one argument"
+        expect { client.request }.to raise_error(ArgumentError, message)
       end
     end
 
-    context "with a single argument (Symbol)" do
-      it "should set the input tag to result in <getUser>" do
+    context "called with a single argument (Symbol)" do
+      it "sets the input tag to result in <getUser>" do
         client.request(:get_user) { soap.input.should == [:getUser, {}] }
       end
 
-      it "should set the target namespace with the default identifier" do
+      it "sets the target namespace with the default identifier" do
         namespace = 'xmlns:wsdl="http://v1_0.ws.auth.order.example.com/"'
         HTTPI::Request.any_instance.expects(:body=).with { |value| value.include? namespace }
 
         client.request :get_user
       end
 
-      it "should not set the target namespace if soap.namespace was set to nil" do
+      it "does not set the target namespace if soap.namespace was set to nil" do
         namespace = 'wsdl="http://v1_0.ws.auth.order.example.com/"'
         HTTPI::Request.any_instance.expects(:body=).with { |value| !value.include?(namespace) }
 
@@ -102,31 +96,31 @@ describe Savon::Client do
       end
     end
 
-    context "with a single argument (String)" do
-      it "should set the input tag to result in <get_user>" do
+    context "called with a single argument (String)" do
+      it "sets the input tag to result in <get_user>" do
         client.request("get_user") { soap.input.should == [:get_user, {}] }
       end
     end
 
-    context "with a Symbol and a Hash" do
-      it "should set the input tag to result in <getUser active='true'>" do
+    context "called with a Symbol and a Hash" do
+      it "sets the input tag to result in <getUser active='true'>" do
         client.request(:get_user, :active => true) { soap.input.should == [:getUser, { :active => true }] }
       end
     end
 
-    context "with two Symbols" do
-      it "should set the input tag to result in <wsdl:getUser>" do
+    context "called with two Symbols" do
+      it "sets the input tag to result in <wsdl:getUser>" do
         client.request(:v1, :get_user) { soap.input.should == [:v1, :getUser, {}] }
       end
 
-      it "should set the target namespace with the given identifier" do
+      it "sets the target namespace with the given identifier" do
         namespace = 'xmlns:v1="http://v1_0.ws.auth.order.example.com/"'
         HTTPI::Request.any_instance.expects(:body=).with { |value| value.include? namespace }
 
         client.request :v1, :get_user
       end
 
-      it "should not set the target namespace if soap.namespace was set to nil" do
+      it "does not set the target namespace if soap.namespace was set to nil" do
         namespace = 'xmlns:v1="http://v1_0.ws.auth.order.example.com/"'
         HTTPI::Request.any_instance.expects(:body=).with { |value| !value.include?(namespace) }
 
@@ -134,95 +128,72 @@ describe Savon::Client do
       end
     end
 
-    context "with two Symbols and a Hash" do
-      it "should set the input tag to result in <wsdl:getUser active='true'>" do
+    context "called with two Symbols and a Hash" do
+      it "sets the input tag to result in <wsdl:getUser active='true'>" do
         client.request(:wsdl, :get_user, :active => true) { soap.input.should == [:wsdl, :getUser, { :active => true }] }
       end
     end
 
-    context "with a block expecting one argument" do
-      it "should yield the SOAP object" do
-        client.request(:authenticate) { |soap| soap.should be_a(Savon::SOAP::XML) }
+    context "called with a block expecting one argument" do
+      it "yields the client instance" do
+        client.request(:authenticate) { |client| client.should be_a(Savon::Client) }
       end
     end
 
-    context "with a block expecting two arguments" do
-      it "should yield the SOAP and WSDL objects" do
-        client.request(:authenticate) do |soap, wsdl|
-          soap.should be_a(Savon::SOAP::XML)
-          wsdl.should be_an(Savon::WSDL::Document)
-        end
-      end
-    end
-
-    context "with a block expecting three arguments" do
-      it "should yield the SOAP, WSDL and HTTP objects" do
-        client.request(:authenticate) do |soap, wsdl, http|
-          soap.should be_a(Savon::SOAP::XML)
-          wsdl.should be_an(Savon::WSDL::Document)
-          http.should be_an(HTTPI::Request)
-        end
-      end
-    end
-
-    context "with a block expecting four arguments" do
-      it "should yield the SOAP, WSDL, HTTP and WSSE objects" do
-        client.request(:authenticate) do |soap, wsdl, http, wsse|
-          soap.should be_a(Savon::SOAP::XML)
-          wsdl.should be_a(Savon::WSDL::Document)
-          http.should be_an(HTTPI::Request)
-          wsse.should be_a(Savon::WSSE)
-        end
-      end
-    end
-
-    context "with a block expecting no arguments" do
-      it "should let you access the SOAP object" do
+    context "called with a block expecting no arguments" do
+      it "lets you access the SOAP object" do
         client.request(:authenticate) { soap.should be_a(Savon::SOAP::XML) }
       end
 
-      it "should let you access the HTTP object" do
+      it "lets you access the HTTP object" do
         client.request(:authenticate) { http.should be_an(HTTPI::Request) }
       end
 
-      it "should let you access the WSSE object" do
+      it "lets you access the WSSE object" do
         client.request(:authenticate) { wsse.should be_a(Savon::WSSE) }
       end
 
-      it "should let you access the WSDL object" do
+      it "lets you access the WSDL object" do
         client.request(:authenticate) { wsdl.should be_a(Savon::WSDL::Document) }
       end
     end
 
-    it "should not set the Cookie header for the next request" do
+    context "called with a block expecting more than one argument" do
+      it "raises an ArgumentError" do
+        message = "Expected a block with an arity of either 0 or 1"
+        expect { client.request(:authenticate) { |one, two| } }.to raise_error(ArgumentError, message)
+      end
+    end
+
+    it "by default does not set the Cookie header for the next request" do
       client.http.headers.expects(:[]=).with("Cookie", anything).never
       client.http.headers.stubs(:[]=).with("SOAPAction", '"authenticate"')
       client.http.headers.stubs(:[]=).with("Content-Type", "text/xml;charset=UTF-8")
 
       client.request :authenticate
     end
-  end
 
-  context "#request with a Set-Cookie response header" do
-    before do
-      HTTPI.stubs(:get).returns(new_response(:body => Fixture.wsdl(:authentication)))
-      HTTPI.stubs(:post).returns(new_response(:headers => { "Set-Cookie" => "some-cookie" }))
-    end
+    context "with a Set-Cookie response header" do
+      before do
+        HTTPI.stubs(:get).returns(new_response(:body => Fixture.wsdl(:authentication)))
+        HTTPI.stubs(:post).returns(new_response(:headers => { "Set-Cookie" => "user:mac" }))
+      end
 
-    it "should set the Cookie header for the next request" do
-      client.http.headers.expects(:[]=).with("Cookie", "some-cookie")
-      client.http.headers.stubs(:[]=).with("SOAPAction", '"authenticate"')
-      client.http.headers.stubs(:[]=).with("Content-Type", "text/xml;charset=UTF-8")
+      it "sets the Cookie header for the next request" do
+        client.http.headers.expects(:[]=).with("Cookie", "user:mac")
+        client.http.headers.stubs(:[]=).with("SOAPAction", '"authenticate"')
+        client.http.headers.stubs(:[]=).with("Content-Type", "text/xml;charset=UTF-8")
 
-      client.request :authenticate
+        client.request :authenticate
+      end
     end
   end
 
   context "with a remote WSDL document" do
-    let(:client) { Savon::Client.new { wsdl.document = Endpoint.wsdl } }
+    let(:client) { Savon::Client.new Endpoint.wsdl }
     before { HTTPI.expects(:get).returns(new_response(:body => Fixture.wsdl(:authentication))) }
 
-    it "should return a list of available SOAP actions" do
+    it "returns a list of available SOAP actions" do
       client.wsdl.soap_actions.should == [:authenticate]
     end
 
@@ -234,7 +205,7 @@ describe Savon::Client do
       end
     end
 
-    it "should execute SOAP requests and return the response" do
+    it "executes SOAP requests and returns the response" do
       HTTPI.expects(:post).returns(new_response)
       response = client.request(:authenticate)
 
@@ -244,11 +215,11 @@ describe Savon::Client do
   end
 
   context "with a local WSDL document" do
-    let(:client) { Savon::Client.new { wsdl.document = "spec/fixtures/wsdl/authentication.xml" } } 
+    let(:client) { Savon::Client.new "spec/fixtures/wsdl/authentication.xml" }
 
     before { HTTPI.expects(:get).never }
 
-    it "should return a list of available SOAP actions" do
+    it "returns a list of available SOAP actions" do
       client.wsdl.soap_actions.should == [:authenticate]
     end
 
@@ -260,14 +231,14 @@ describe Savon::Client do
       end
     end
 
-    it "should get #element_form_default from the WSDL" do
+    it "gets the value of #element_form_default from the WSDL" do
       HTTPI.stubs(:post).returns(new_response)
       Savon::WSDL::Document.any_instance.expects(:element_form_default).returns(:qualified)
 
       client.request :authenticate
     end
 
-    it "should execute SOAP requests and return the response" do
+    it "executes SOAP requests and returns the response" do
       HTTPI.expects(:post).returns(new_response)
       response = client.request(:authenticate)
 
@@ -344,8 +315,8 @@ describe Savon::Client do
 
     before { HTTPI.expects(:get).never }
 
-    it "raise an ArgumentError when trying to access the WSDL" do
-      lambda { client.wsdl.soap_actions }.should raise_error(ArgumentError)
+    it "raises an ArgumentError when trying to access the WSDL" do
+      expect { client.wsdl.soap_actions }.to raise_error(ArgumentError)
     end
 
     it "adds a SOAPAction header containing the SOAP action name" do
@@ -356,14 +327,14 @@ describe Savon::Client do
       end
     end
 
-    it "should not get #element_form_default from the WSDL" do
+    it "does not try to get the value of #element_form_default from the WSDL" do
       HTTPI.stubs(:post).returns(new_response)
       Savon::WSDL::Document.any_instance.expects(:element_form_default).never
 
       client.request :authenticate
     end
 
-    it "should execute SOAP requests and return the response" do
+    it "executes SOAP requests and returns the response" do
       HTTPI.expects(:post).returns(new_response)
       response = client.request(:authenticate)
 
@@ -380,10 +351,13 @@ describe Savon::Client do
       end
     end
 
-    before { HTTPI::expects(:post).returns(new_response(:code => 500, :body => Fixture.response(:soap_fault))) }
+    before do
+      response = new_response :code => 500, :body => Fixture.response(:soap_fault)
+      HTTPI::expects(:post).returns(response)
+    end
 
-    it "should raise a Savon::SOAP::Fault" do
-      lambda { client.request :authenticate }.should raise_error(Savon::SOAP::Fault)
+    it "raises a Savon::SOAP::Fault" do
+      expect { client.request :authenticate }.to raise_error(Savon::SOAP::Fault)
     end
   end
 
@@ -397,8 +371,8 @@ describe Savon::Client do
 
     before { HTTPI::expects(:post).returns(new_response(:code => 500)) }
 
-    it "should raise a Savon::HTTP::Error" do
-      lambda { client.request :authenticate }.should raise_error(Savon::HTTP::Error)
+    it "raises a Savon::HTTP::Error" do
+      expect { client.request :authenticate }.to raise_error(Savon::HTTP::Error)
     end
   end